Extract `write_table` from `lib_write` into `lib_tables`
