+++ fractal_tasks_core/tasks/napari_workflows_wrapper.py
@@ -37,9 +37,9 @@ from fractal_tasks_core.lib_regions_of_interest import (
     convert_ROI_table_to_indices,
 )
 from fractal_tasks_core.lib_regions_of_interest import load_region
+from fractal_tasks_core.lib_tables import write_table
 from fractal_tasks_core.lib_upscale_array import upscale_array
 from fractal_tasks_core.lib_write import prepare_label_group
-from fractal_tasks_core.lib_write import write_table
 from fractal_tasks_core.lib_zattrs_utils import rescale_datasets
